Remove container only if existing

+++ rakelib/helper/docker.rb
@@ -24,7 +24,7 @@ module DOCKER
     name = "test-server-#{server}"
     image = SETUP::WORKBOOK[server][:docker][:image]
     image_id = `docker images -q #{image}`
-    system "docker rm #{name}"
+    system "docker rm #{name}" if container_exists?(server)
     system "docker rmi #{image_id}" unless image_id.empty?
   end
